added clones for 2 CNetClient functions
@elexis1

I think this is very close to what you wanted me to do.

The result now is that clicking "host game" results
in the dialog box "connecting to server".

Looking at NetServer, I think what's needed now to get past that
is code similar to this in RunStep() but for GuiPoll() and
SendGameSetupMessage()

`
GetScriptInterface().ParseJSON(newGameAttributes.back(),
&gameAttributesVal);

UpdateGameAttributes(&gameAttributesVal);
`

I didn't know how to try that though.